基础点赞、收藏、评论功能

+++ redbook-service/redbook-service-comment/src/main/java/com/itcast/controller/CommentController.java
@@ -31,14 +31,14 @@ public class CommentController {
     @Operation(summary = "获取评论列表", description = "获取指定笔记的所有评论")
     @GetMapping("/getCommentList/{noteId}")
     public Result<List<CommentVo>> getCommentList(
-            @Parameter(description = "笔记ID", required = true) @PathVariable("noteId") Integer noteId) throws ParseException {
+            @Parameter(description = "笔记ID", required = true) @PathVariable("noteId") Long noteId) throws ParseException {
         return commentService.getCommentList(noteId);
     }
 
     @Operation(summary = "获取评论数量", description = "获取指定笔记的评论总数")
     @GetMapping("/getCommentCount/{noteId}")
     public Result<Integer> getCommentCount(
-            @Parameter(description = "笔记ID", required = true) @PathVariable("noteId") Integer noteId) {
+            @Parameter(description = "笔记ID", required = true) @PathVariable("noteId") Long noteId) {
         return commentService.getCommentCount(noteId);
     }
