+++ redbook-service/redbook-service-note/src/main/java/com/itcast/model/pojo/Collection.java
@@ -23,7 +23,7 @@ public class Collection implements Serializable {
      * 笔记id
      */
     @TableField("note_id")
-    private Integer noteId;
+    private Long noteId;
 
     /**
      * 用户id
