+++ redbook-service/redbook-service-note/src/main/java/com/itcast/service/CollectionService.java
@@ -5,5 +5,5 @@ import com.itcast.result.Result;
 public interface CollectionService {
     Result<Boolean> isCollection(Long noteId);
 
-    Result<Void> collection(Integer noteId);
+    Result<Void> collection(Long noteId);
 }
